git-svn-id: https://gsphere.genoscope.cns.fr/svn/NGL/trunk/ngl@4369 d0e871a8-0b98-4e59-9d77-5ea3c331444f

+++ lib-ngl-common/app/models/laboratory/reporting/instance/Column.java
@@ -27,7 +27,7 @@ public class Column implements IValidation{
 	
 	
 	public String[] modes; //chart or table or the two
-	public Float position;
+	public Double position;
 	
 	public String[] queryIncludeKeys; //if void used property
